requirements.txt
# Core dependencies for Supply Chain Time Series Forecasting
numpy==1.26.4
pandas==2.2.0
openpyxl==3.1.2

# Time series modeling
statsmodels==0.14.1

# Machine learning utilities
scikit-learn==1.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Configuration and utilities
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2
python-dateutil==2.8.2
typing-extensions==4.9.0
